Primero se hace que importacion de tabla de dominio se haga en contenedor en vez de fullScreenDialog
